Markprompt lint warnings
