+++ packages/react/src/Markprompt.test.tsx
@@ -2,7 +2,8 @@ import { act, render, screen, waitFor } from '@testing-library/react';
 import { userEvent } from '@testing-library/user-event';
 import { describe, expect, it, vi } from 'vitest';
 
-import { Markprompt, closeMarkprompt, openMarkprompt } from './index.js';
+import { Markprompt } from './index.js';
+import { closeMarkprompt, openMarkprompt } from './utils.js';
 
 describe('Markprompt', () => {
   it('renders', () => {
